+++ src/main/java/ie/ul/cs4227/Bass/Service/ITopUpService.java
@@ -6,7 +6,7 @@ import ie.ul.cs4227.Bass.Entity.Recharge;
 import ie.ul.cs4227.Bass.Entity.User;
 
 public interface ITopUpService {
- public abstract Boolean TopUp(User u,Integer amount,String type) ;
+ public abstract Boolean TopUp(User u,Float amount,String type) ;
  public abstract ArrayList<Recharge> getTopUpList() ;
  public abstract Recharge selectTopUpById(Integer tId) ;
 }
